refactor: move ImageUpload presentation to @app/form-component
Move the ImageUpload component's presentational logic from app-frontend
(src/layout/ImageUpload) into the shared @app/form-component lib as
ImageUploadLayout, leaving a thin data/update wrapper behind.

The lib owns all presentation: the cropper orchestration, canvas
(draw/drag/keyboard/zoom), controllers, dropzone, preview, focus hooks,
file-upload/crop-save logic and the geometry utils. It is fully controlled:
the wrapper injects storedImage/imageUrl and commits via onSave(file)/onDelete()
callbacks. Validation + grid are wired via ComponentStructure inside the lib
(no ComponentStructureWrapper in app-frontend).

The wrapper keeps only runtime wiring: useImageFile (attachments), the
allowedContentTypes guard, useLabelData and useComponentStructureData.

- Add useIsMobileOrTablet to the lib's useDeviceWidths hook (used by the dropzone).
- Split imageUploadUtils: pure geometry/zoom helpers move to the lib (with
  lib-owned CropConfig/StoredImage types); the app keeps only
  isAllowedContentTypesValid (app-metadata guard).
- Add Storybook docs page (stories + mdx) and unit tests in the lib; keep the
  app integration tests as the source of truth for preserved behaviour.
