+++ app-libs/form-component/src/app-components/hooks/index.ts
@@ -2,6 +2,7 @@ export {
   useIsMini,
   useIsMobile,
   useIsTablet,
+  useIsMobileOrTablet,
   useBrowserWidth,
   breakpoints,
 } from './useDeviceWidths';
