+++ app-libs/form-component/src/layout-components/ImageUpload/ImageCanvas/hooks/useDragInteraction.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 
-import type { Position } from 'src/layout/ImageUpload/imageUploadUtils';
+import type { Position } from '@app/form-component/layout-components/ImageUpload/imageUploadUtils';
 
 type UseDragInteractionProps = {
   canvasRef: React.RefObject<HTMLCanvasElement | null>;
@@ -8,7 +8,11 @@ type UseDragInteractionProps = {
   onPositionChange: (newPosition: Position) => void;
 };
 
-export const useDragInteraction = ({ canvasRef, position, onPositionChange }: UseDragInteractionProps) => {
+export const useDragInteraction = ({
+  canvasRef,
+  position,
+  onPositionChange,
+}: UseDragInteractionProps) => {
   const handlePointerDown = (e: React.PointerEvent<HTMLCanvasElement>) => {
     e.preventDefault();
