+++ app-libs/form-component/src/layout-components/ImageUpload/ImageUploadLayout.mdx
@@ -1,0 +1,9 @@
+import { Meta } from '@storybook/addon-docs/blocks';
+
+import { LayoutComponentDocs } from '../common/storybook';
+import * as ImageUploadStories from './ImageUploadLayout.stories';
+import { IMAGE_UPLOAD_PROP_CATEGORIES } from './ImageUploadLayout.stories';
+
+<Meta of={ImageUploadStories} />
+
+<LayoutComponentDocs categories={IMAGE_UPLOAD_PROP_CATEGORIES} />
