+++ app-libs/form-component/src/layout-components/ImageUpload/hooks/useFocusWhenRemoved.tsx
@@ -1,7 +1,10 @@
 import { useEffect, useRef } from 'react';
 import type React from 'react';
 
-export function useFocusWhenRemoved<T>(value: T | null | undefined, elementRef: React.RefObject<HTMLElement | null>) {
+export function useFocusWhenRemoved<T>(
+  value: T | null | undefined,
+  elementRef: React.RefObject<HTMLElement | null>,
+) {
   const prevValueRef = useRef<T | null | undefined>(value);
 
   useEffect(() => {
