--- src/App/frontend/src/layout/ImageUpload/ImageDropzone.tsx
@@ -1,61 +1,0 @@
-import React from 'react';
-
-import { Dropzone, getDescriptionId } from '@app/form-component';
-import cn from 'classnames';
-import type { IDropzoneProps } from '@app/form-component';
-
-import { Lang } from 'src/features/language/Lang';
-import { useIsMobileOrTablet } from 'src/hooks/useDeviceWidths';
-import classes from 'src/layout/ImageUpload/ImageDropzone.module.css';
-
-type ImageDropzoneProps = {
-  baseComponentId: string;
-  hasErrors: boolean;
-  readOnly: boolean;
-  dropzoneInputRef?: React.RefObject<HTMLInputElement | null>;
-} & Pick<IDropzoneProps, 'onDrop'>;
-
-export function ImageDropzone({ baseComponentId, hasErrors, readOnly, onDrop, dropzoneInputRef }: ImageDropzoneProps) {
-  const [dragActive, setDragActive] = React.useState(false);
-  const isMobile = useIsMobileOrTablet();
-  const descriptionId = getDescriptionId(baseComponentId);
-  const dragLabelId = `file-upload-drag-${baseComponentId}`;
-  const formatLabelId = `file-upload-format-${baseComponentId}`;
-  const ariaDescribedBy = [descriptionId, dragLabelId, formatLabelId].filter(Boolean).join(' ');
-
-  return (
-    <Dropzone
-      id={baseComponentId}
-      readOnly={readOnly}
-      onDrop={onDrop}
-      onDragActiveChange={setDragActive}
-      hasValidationMessages={hasErrors}
-      acceptedFiles={{ 'image/*': [] }}
-      data-color='neutral'
-      className={cn(classes.placeholder, { [classes.dragActive]: dragActive })}
-      describedBy={ariaDescribedBy}
-      inputRef={dropzoneInputRef}
-    >
-      <div className={classes.dropZone}>
-        <b id={dragLabelId}>
-          {isMobile ? (
-            <span className={classes.underLine}>
-              <Lang id='form_filler.file_uploader_upload' />
-            </span>
-          ) : (
-            <>
-              <Lang id='form_filler.file_uploader_drag' />
-              <span className={classes.underLine}>
-                {' '}
-                <Lang id='form_filler.file_uploader_find' />
-              </span>
-            </>
-          )}
-        </b>
-        <span id={formatLabelId}>
-          <Lang id='image_upload_component.valid_file_types' />
-        </span>
-      </div>
-    </Dropzone>
-  );
-}
